I should really do this more often.
I'm almost to milestone 1, but I might need to finish it over the break. This commit is finishes up most of the movement rules, and also gets it displaying the board in the console. I still need to finish the castling rules and get the program to a point where you can input moves in the console. 

I also started some of the framework for the AI player.