CASTLING
Finally got castling working more or less perfectly. I got a stack overflow from what looked like an infinite loop at one point, but I can't seem to reproduce it.

At this point, almost all of the UI work is done. I still need to make it so that the players have to take turns, though. After that, I'll make it so the game ends on a checkmate.